requirements.txt
gitpython>=3.1.40
openai>=1.50.0
slack-sdk>=3.26.1
python-dotenv>=1.0.0
click>=8.1.7
pydantic>=2.10.0
rich>=13.7.0
